delivery: Allow making additions to the message header on per-recipient basis
Received and Delivered-To headers are supposed to be inserted with
recipient-specific information.

The update to allow this behavior required major changes and to
preserve performance Delivery.BodyParsed method now closely resembles
maddy interface for body buffering.

Well, being honest, this whole Delivery interface is designed for
tight integration with maddy.