Prefetch mboxId for INBOX when looking up user creds
This slightly improves performance (by saving one query roundtrip)
for message additions and for SELECT INBOX.

On the other hand, it will break if INBOX is renamed between
GetMailbox and GetUser. This is assumed to be highly unlikely
and thus is not handled safetly. Mailbox IDs are not reused, so
this will not accidentely give access to someones else mailbox.